AIStudio/requirements.txt
# AIStudio Python脚本依赖
# 核心依赖
openai>=1.0.0
python-pptx>=0.6.21
requests>=2.25.0
moviepy>=1.0.3

# 增强功能
Pillow>=8.0.0

# 可选功能 (根据Python版本自动选择)
TTS>=0.22.0  # Python 3.10-3.11
TTS==0.21.3  # Python 3.12

# 云端TTS服务
dashscope>=1.17.0  # 阿里云DashScope SDK (Qwen TTS)
